+++ index.js
@@ -12,6 +12,7 @@ module.exports = function(homebridge) {
     homebridge.registerAccessory("homebridge-http-webhooks", "HttpWebHookSensor", HttpWebHookSensorAccessory);
     homebridge.registerAccessory("homebridge-http-webhooks", "HttpWebHookSwitch", HttpWebHookSwitchAccessory);
     homebridge.registerAccessory("homebridge-http-webhooks", "HttpWebHookDoor", HttpWebHookDoorAccessory);
+    homebridge.registerAccessory("homebridge-http-webhooks", "HttpWebHookPushButton", HttpWebHookPushButtonAccessory);
 };
 
 function HttpWebHooksPlatform(log, config){
@@ -21,6 +22,7 @@ function HttpWebHooksPlatform(log, config){
     this.sensors = config["sensors"] || [];
     this.switches = config["switches"] || [];
     this.doors = config["doors"] || [];
+    this.pushButtons = config["pushbuttons"] || [];
     this.storage = require('node-persist');
     this.storage.initSync({dir:this.cacheDirectory});
 }
@@ -43,6 +45,11 @@ HttpWebHooksPlatform.prototype = {
         for(i = 0; i < this.doors.length; i++){
             var doorAccessory = new HttpWebHookDoorAccessory(this.log, this.doors[i], this.storage);
             accessories.push(doorAccessory);
+        }
+
+        for(i = 0; i < this.pushButtons.length; i++){
+            var pushButtonsAccessory = new HttpWebHookPushButtonAccessory(this.log, this.pushButtons[i], this.storage);
+            accessories.push(pushButtonsAccessory);
         }
 
         var accessoriesCount = accessories.length;
@@ -234,6 +241,9 @@ HttpWebHookSwitchAccessory.prototype.setState = function(powerOn, callback) {
                 callback(err || new Error("Request to '"+url+"' was not succesful."));
             }
         }).bind(this));
+    }
+    else {
+        callback(null);
     }
 };
 
@@ -440,4 +450,67 @@ HttpWebHookDoorAccessory.prototype.setState = function(state, callback) {
 
 HttpWebHookDoorAccessory.prototype.getServices = function() {
     return [this.service];
+};
+
+
+
+function HttpWebHookPushButtonAccessory(log, pushButtonConfig, storage) {
+    this.log = log;
+    this.id = pushButtonConfig["id"];
+    this.name = pushButtonConfig["name"];
+    this.pushURL = pushButtonConfig["push_url"] || "";
+
+    this.service = new Service.Switch(this.name);
+    this.changeHandler = (function(newState) {
+        if(newState) {
+            this.log("Change HomeKit state for push button to '%s'.", newState);
+            this.service.getCharacteristic(Characteristic.On)
+                    .setValue(newState, undefined, 'fromHTTPWebhooks');
+        }
+    }).bind(this);
+    this.service
+        .getCharacteristic(Characteristic.On)
+        .on('get', this.getState.bind(this))
+        .on('set', this.setState.bind(this));
+}
+
+HttpWebHookPushButtonAccessory.prototype.getState = function(callback) {
+    this.log("Getting current state for '%s'...", this.id);
+    var state = false;
+    callback(null, state);
+};
+
+HttpWebHookPushButtonAccessory.prototype.setState = function(powerOn, callback) {
+    this.log("Push buttons state change for '%s'...", this.id);
+    if(!powerOn) {
+        callback(null);
+    }
+    else if(this.pushURL === "") {
+        callback(null);
+        setTimeout(function() {
+            this.service.getCharacteristic(Characteristic.On).setValue(false, undefined, 'fromTimeoutCall');
+        }.bind(this), 1000);
+    }
+    else {
+        request.get({
+            url: this.pushURL,
+            timeout: DEFAULT_REQUEST_TIMEOUT
+        }, (function(err, response, body) {
+            var statusCode = response && response.statusCode ? response.statusCode: -1;
+            this.log("Request to '%s' finished with status code '%s' and body '%s'.", url, statusCode, body, err);
+            if (!err && statusCode == 200) {
+                callback(null);
+            }
+            else {
+                callback(err || new Error("Request to '"+url+"' was not succesful."));
+            }
+            setTimeout(function() {
+                this.service.getCharacteristic(Characteristic.On).setValue(false, undefined, 'fromTimeoutCall');
+            }.bind(this), 1000);
+        }).bind(this));
+    }
+};
+
+HttpWebHookPushButtonAccessory.prototype.getServices = function() {
+  return [this.service];
 };